backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.136.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.38",
]
